+++ src/main/java/be/tempsdor/tempsdor/services/RoleService.java
@@ -6,6 +6,6 @@ import be.tempsdor.tempsdor.exceptions.ElementsNotFoundException;
 
 import java.util.List;
 
-public interface RoleService extends CrudService<RoleDTO, RoleDTO, Integer> {
+public interface RoleService extends CrudService<RoleDTO, RoleDTO, Long> {
     RoleDTO getByName(String name) throws ElementNotFoundException;
 }
